requirements.txt
apscheduler==3.10.4
authlib<1.0.0
azure-identity==1.12.0
azure-storage-blob==12.14.1
boto3==1.26.67
chardet==5.1.0
elastic-apm[flask]
elasticsearch==8.6.1
flask==2.2.5
flask_caching==2.0.2
gevent==23.9.1
gunicorn==20.1.0
hogwarts-auth==2.0.3.14851
hogwarts-spellbook==0.0.1.15222
netifaces==0.11.0
packaging==23.0
passlib==1.7.4
prometheus-client==0.17.1
pyjwt==2.6.0
pyroute2.core==0.6.13
pysftp==0.2.9
python-baseconv==1.2.2
python-datemath==1.5.5
PyYAML==6.0
redis==4.5.4
requests==2.31.0
wsproto==1.2.0
hogwarts-auth==2.0.3.14851
chevron==0.14.0
